requirements.txt
numpy~=1.24.1
tqdm~=4.64.1
pandas~=1.5.2
matplotlib~=3.6.2
seaborn~=0.12.1
pysam~=0.20.0
scikit-learn~=1.2.0
bio~=1.5.3
torch~=1.13.1
transformers~=4.25.1
tensorflow~=2.11.0
PyYAML~=6.0